refactor: extract QuoteApiErrorButton pure component
